APIで汎用的なのはApiHelperに定義するようにした
