+++ app/api/v1/root.rb
@@ -2,6 +2,7 @@
 
 module V1
   class Root < Grape::API
+    helpers Api_Helpers
     version :v1
     format :json
