+++ app/api/v1/self_care_classifications.rb
@@ -37,23 +37,6 @@ module V1
           update_params
         end
 
-        # 1人のユーザーしか使わない想定
-        def current_user
-          User.last
-        end
-
-        def create_error_message_from_model(model)
-          error_message = model.errors.messages.reduce('') do |message, (name, msgs)|
-            message += "#{name}:"
-            msgs.each do |msg|
-              message += "#{msg},"
-            end
-            message.slice!(-1)
-            message += "\n"
-          end
-          error_message.slice!(-1)
-          error_message
-        end
       end
 
       desc 'セルフケア分類をまとめて設定する 戻り値ResultResponse'
